# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Codalla AI Coding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Success Criteria ────────────────────────────────────────────────────────
...
  /github/repos:
    get:
      operationId: listGithubRepos
      tags: [github]
      summary: List GitHub repos for the authenticated user
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: perPage
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of repos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GithubRepo"
  /projects/{projectId}/github/clone:
    post:
      operationId: cloneRepo
      tags: [github]
      summary: Clone a GitHub repo into a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloneInput"
      responses:
        "200":
          description: Clone result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitResult"
  /projects/{projectId}/github/status:
    get:
      operationId: getGitStatus
      tags: [github]
      summary: Get git status (modified, added, deleted files)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Git status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
  /git-diff:
    get:
      operationId: getFileDiff
      tags: [github]
      summary: Get diff for a specific file
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: filePath
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File diff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDiff"
  /projects/{projectId}/github/commit:
    post:
      operationId: commitAndPush
      tags: [github]
      summary: Stage, commit, and push changes
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitInput"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitResult"
  /projects/{projectId}/github/pull:
    post:
      operationId: pullRepo
      tags: [github]
      summary: Pull latest changes from remote
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullInput"
      responses:
        "200":
          description: Pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitResult"
  /projects/{projectId}/github/branches:
    get:
      operationId: listBranches
      tags: [github]
      summary: List branches in the repo
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchList"
  /projects/{projectId}/github/checkout:
    post:
      operationId: checkoutBranch
      tags: [github]
      summary: Checkout a branch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitResult"

  # ─── AI Models ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, localPath, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        isShared:
          type: boolean
        isOwner:
          type: boolean
        localPath:
          type: string
        gitRemoteUrl:
          type: ["string", "null"]
        currentBranch:
          type: string
        lastSynced:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        story:
          type: ["string", "null"]
        target:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isShared:
          type: boolean
        gitRemoteUrl:
          type: string
        currentBranch:
          type: string
        story:
          type: string
        target:
          type: string

    # ─── Success Criteria ──────────────────────────────────────────────────────
...
    GithubRepo:
      type: object
      required: [id, name, fullName, cloneUrl, private, defaultBranch]
      properties:
        id:
          type: integer
        name:
          type: string
        fullName:
          type: string
        description:
          type: ["string", "null"]
        cloneUrl:
          type: string
        sshUrl:
          type: string
        private:
          type: boolean
        defaultBranch:
          type: string
        language:
          type: ["string", "null"]
        stargazersCount:
          type: integer
        updatedAt:
          type: string
    CloneInput:
      type: object
      required: [repoUrl]
      properties:
        repoUrl:
          type: string
        branch:
          type: string
        token:
          type: string
    CommitInput:
      type: object
      required: [message, token]
      properties:
        message:
          type: string
        token:
          type: string
        files:
          type: array
          items:
            type: string
    PullInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    CheckoutInput:
      type: object
      required: [branch]
      properties:
        branch:
          type: string
        create:
          type: boolean
    GitResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        commitHash:
          type: ["string", "null"]
    GitStatus:
      type: object
      required: [modified, added, deleted, untracked]
      properties:
        modified:
          type: array
          items:
            type: string
        added:
          type: array
          items:
            type: string
        deleted:
          type: array
          items:
            type: string
        untracked:
          type: array
          items:
            type: string
        branch:
          type: string
    FileDiff:
      type: object
      required: [path, diff]
      properties:
        path:
          type: string
        diff:
          type: string
        additions:
          type: integer
        deletions:
          type: integer
    BranchList:
      type: object
      required: [branches, current]
      properties:
        branches:
          type: array
          items:
            type: string
        current:
          type: string

    # ─── AI ────────────────────────────────────────────────────────────────────
...